New folder (2)/requirements.txt
requests>=2.31.0
newspaper3k>=0.2.8
google-generativeai>=0.3.0
pillow>=10.0.0
lxml>=4.9.0
lxml_html_clean>=0.4.0
beautifulsoup4>=4.12.0

